Expand Series-09 documentation and refactor examples
Added detailed explanations and examples to ReadMe.md, enhancing the
understanding of Node.js Event Loop in Series-09. The package.json and
package-lock.json were updated, reflecting the correct series and
dependencies. Removed obsolete example files, indicating a shift towards
a more structured and illustrative codebase with added new examples
demonstrating the use of timers and event queue handling in Node.js. The
main.ts now points to the relocated source files within `src/examples`.
This comprehensive update consolidates crucial concepts, making the
learning material more accessible and organized for better user
engagement.

Related to the documentation overhaul for the TypeScript Node series.
