--- series-09/examples/diff-set-timeout-and-set-interval.ts
@@ -1,10 +1,0 @@
-let i = 0;
-
-function increment() {
-  console.log(++i);
-  if (i <= 10) {
-    setTimeout(increment, 50);
-  }
-}
-
-increment();
