--- series-09/main.ts
@@ -1,5 +1,0 @@
-// import "./examples/set-timeout";
-// import "./examples/set-interval";
-// import "./examples/diff-set-timeout-and-set-interval";
-// import "./examples/set-timeout-and-set-immediate";
-import "./examples/check-queue";
