+++ series-09/src/examples/setIntervalExample.ts
