+++ series-09/src/examples/setTimeoutExample.ts
@@ -1,5 +1,6 @@
 console.time("setTimeout");
 
+// 100 milliseconds is not grantee it is the time after that it can pop from queue
 setTimeout(() => {
   console.log("Timer went off");
   console.timeEnd("setTimeout");
