+++ series-09/src/main.ts
@@ -1,0 +1,5 @@
+// import "./examples/setTimeoutExample";
+// import "./examples/setIntervalExample";
+// import "./examples/overrideSetIntervalWithSetTimeoutExample";
+// import "./examples/setImmediateAndSetTimeoutExample";
+import "./examples/checkQueueExample";
